3905: Do not retrieve online urls to determine if object is online
Before the search abstraction layer was introduced the getOnline_url()
method accepted a boolean to determine whether to retrieve relations
for fetching online urls for an object or not. The availability field 
Formatter was the only element which used this functionality to prevent 
relations from being fetched when determining whether a material was 
online or not.

When the search abstraction layer was introduced a new method for
determining if an object is online without fetching relations was 
introduced: isOnline() and getOnline_url() lost the boolean and always
fetched relations. However the availability module was not updated
accordingly. Consequently display of the availability field would
always fetch relations which potentially could lead to a new search
for each object being displayed.

With an empty cache this means that performing a search leads to
a search request + requests per object.

In the context of the availability field formatter the online urls are
still not needed so we can replace the call to getOnline_url() with
isOnline() and thus prevent a lot of unneeded requests.