Initial work on React.js user dashboard frontend component
This is the result of creating a new module, running create-react-app
to create a skeleton and from the building an initial set of
components for showing user loan and reservation status.

Running the application will display a development version of the
component with mock data. Styling is also a work in progress.

The component is not tied to Drupal in any way at the moment apart
from residing inside a module folder.