Add support for specifying an API token using a data attribute
We assume that the server will be responsible for acquiring a token
for accessing the OpenPlatform on behalf of the user. This way we
avoid exposing our client id and secret to the frontend.

To pass the token to the frontend we let the server generate it as a
date attribute.