Implement translation handling using Drupal translations
We want to allow editors to manage translation using the Drupal
workflow they already know for translating server side strings
Drupal supports clientside translations using Drupal.t() but the
target strings need to be exposed client side for this to work.

Drupal handles this by parsing included JavaScript files for use of
Drupal.t() in _locale_parse_js_file(). This does not work with 
JavaScript files generated by Webpack.

Consequently we introduce a webpack plugin which parses files for
strings used with Drupal.t() and generates a corresponding parsable
file.

This allows us to use Drupal translations in our components.

For the plugin to work we have to call Drupal.t() directly in our
component code. This creates a coupling to Drupal which may not be
desirable but we think is acceptable.

Code which does not have Drupal.t() available can easily provide a
polyfill. If we wanted to avoid the direct coupling to Drupal
entirely we could update the webpack plugin to look for other
patterns and expose these as translation strings and then use the
other pattern to call Drupal.t() indirectly.