Require newest Chrome Mink Driver for Behat tests
The latest version of Chrome requires specific headers to be set
for remote debugging to work. This has been addressed in the latest
version of the package so update it accordingly.

This may not be a problem in our CI environments at the moment but
could be in the future. It is a problem when trying to use a local
version of Chrome.

See https://gitlab.com/DMore/chrome-mink-driver/merge_requests/63.