Use more recent build of ddb-react
We have updated to the library which may not be available in the
deployed profile. When working as a admin uploaded module we cannot
update this version so we have to provide our own.

By using libraries_info_alter we can set a custom path to the library
so this gets loaded instead.