Rework presentation of collapsed objects
Instead of showing compact version of materials within a title group
we hide them all by default and instead add a button to the first
material of the group. Clicking the button will toggle the other 
materials within the group. 

This requires a rework of logic, behavior and styling as we need the
groups to be available at all levels.