4793: Strip certain characters from keys before performing DB search
Some characters may be valid CQL and thus work when searching for
materials but cause problems when used for CMS content search. One 
example of such a character is ‘. When included in the search keys
a database error will occur such as:

SQLSTATE[42000]: Syntax error or access violation: 1064 You have an 
error in your SQL syntax; check the manual that corresponds to your 
MySQL server version for the right syntax to use near '%' AS w0, 
t.word LIKE '%potter'%' AS w1 FROM 
search_api_db_default_multiple_ind' at line 1

Tracing this error goes deep into the Drupal database layer and 
without a proper fix.

The simplest way to address this seems to be to just strip the
characters.