Prevent multiple confirmation dialogs when ordering
The state of our order material app is controlled by the pid(s) it 
will order when the button is clicked. If multiple apps are 
configured orders the same pids the apps will have the same state. In
the finished state the apps show a success dialog. If multiple apps
end up in the finished state then multiple dialogs are shown.

To avoid this we only configure the first object to have pids for all
the materials which are considered equivalent. Consequently the
app for the first material will place and order for any of equivalent
material while the rest will only order the specific one.

Since the apps then work with different pids they will also have
different state. Ordering using one app will not be considered as if
all apps have been used to place an order.

Consequently only one confirmation dialog will be shown.