5206: Base detection of local items on availability
Before we based this on the presence of the agency in the id. However
this is not a reliable method. Items can have a local id without
in fact having any local physical objects associated which can be
loaned. Also items can have a basis id but still have local physical
objects associated.

Thus we have to base our detection of local objects entirely on
availability.