5206: Only prefer basis objects to local objects on collection view
If the preference is activated on the search page then the list of
material types disappears. This is most likely due to the heavyhanded
approach taken during the filtering.

To avoid this we only filter on the collection view. This may be
similarly heavy handed but this is the territory we are in.